Updated so that analysis can be split into multiple smaller jobs, temporal derivative as regressor now an option for resting state, load templates locally, bug fixes

+++ multiverse/batch_multiverse.py
@@ -9,9 +9,8 @@ import pickle
 import sys
 import shutil
 import os
-import getpass
-import subprocess
-from functions import organize
+import glob
+from nipype import config
 
 processed = '/scratch/processed'
     
@@ -20,6 +19,9 @@ if len(sys.argv) > 3:
     task = sys.argv[2]
     profile = sys.argv[3]
     
+save_dir = processed + '/checkpoints_' + task + '_batch_' + str(batch)
+config.set("execution", "crashdump_dir", save_dir)
+    
 working_dir = '/scratch/{task}_working_dir_{batch}'.format(task=task, batch=batch)
 
 with open(processed + '/reproducibility/' + task + '_workflow_' + batch + '.pkl', 'rb') as wf:
@@ -27,15 +29,11 @@ with open(processed + '/reproducibility/' + task + '_workflow_' + batch + '.pkl'
     
 workflow.run(plugin='IPython', plugin_args={'profile': profile, 'task': task, 'batch': batch})
 
-if os.path.exists(working_dir):
+with open('/code/multiverse/configuration/general_configuration.pkl', 'rb') as f:
+    configuration = pickle.load(f)
+
+if os.path.exists(save_dir) and not glob.glob(save_dir + '/crash-*'):
+    os.rename(save_dir, save_dir + '_done')
+
+if os.path.exists(working_dir) and not configuration['debug'] and not glob.glob(save_dir + '/crash-*'):
     shutil.rmtree(working_dir)
-
-user = getpass.getuser()
-sq = str(subprocess.check_output(['squeue', '--user={0}'.format(user)]))
-out_frame = processed + '/{task}.pkl'.format(task=task)
-
-if sq.count('batch.sh') == 1:
-    paths = organize(task, out_frame)
-    
-    #INSERT DATA PROCESSING HERE - TO BE WRITTEN ONCE WE HAVE TEST DATA
-
